Move zeta helpers into their own modules
